Remove si-aws-ec2 sipkg for testing
This commit removes the "si-aws-ec2" sipkg for testing. It is from Sept
2023 and is outdated. Not only that, but we have since shifted to test
exclusive schemas written and maintained in Rust. Removing this package
and refactoring tests that use it should result in a signficant speedup
in migrations when running tests.

This commit also breaks up the "component" module for integration tests
by splitting out "delete" tests into their own module. Some tests in
the new module have been renamed for standardization purposes.
Otherwise, the tests have remained untouched.
